requirements.txt
fastapi
gunicorn
psycopg2-binary